pkdata/requirements.txt
sqlalchemy
python-binance
btfxwss
fcoin
pymysql
requests
